Update method getColumnHeaders()
This method is more "clever" now, in that it won't go pulling in the next line from the CSV file when it is invoked. This alleviates the danger of pulling a row of "values" rather than the expected column headers.

getColumnHeaders() is now public, so clients can get the column headers on demand.

Main.java test file has been updated to test this method